Move cdk.json to infrastructure folder
This fixes an issue where the cdk.out folder would be created in the infrastructure folder, but expected to be in the root folder (where cdk deploy was executed from).
By moving the cdk.json to the infrastructure folder and executing cdk deploy from there, the cdk.out folder is created where it is expected.